Updated: Meeting Table, Created (routes): Create Meeting and Edit Agenda
The Meeting Table component has been updated - handleClickRow was added to every table row - when a user clicks on some row they are redirected to Edit Agenda (todo: depending on whether the agenda is finalized or not, or there is a meeting protocol for a specific meeting, the user should be redirected accordingly. for now it only leads to EditAgenda).

Edit Agenda route has been created - it's a table where all cells/inputs are editable, and when user clicks on Save all the content is displayed in console

Create meeting component is basically a form with modal components for all the info needed for a create meeting. Save meeting ->console.log